+++ src/swade.ts
@@ -21,7 +21,11 @@ import SwadeCharacterSheet from './module/sheets/SwadeCharacterSheet';
 import SwadeItemSheet from './module/sheets/SwadeItemSheet';
 import SwadeNPCSheet from './module/sheets/SwadeNPCSheet';
 import SwadeVehicleSheet from './module/sheets/SwadeVehicleSheet';
-import { rollInitiative, _sortCombatants } from './module/SwadeCombat';
+import {
+  rollInitiative,
+  _sortCombatants,
+  resetAll,
+} from './module/SwadeCombat';
 import SwadeHooks from './module/SwadeHooks';
 import { SwadeSocketHandler } from './module/SwadeSocketHandler';
 import { rollPowerMacro, rollSkillMacro, rollWeaponMacro } from './module/util';
@@ -54,6 +58,7 @@ Hooks.once('init', () => {
   //Overwrite method prototypes
   Combat.prototype.rollInitiative = rollInitiative;
   Combat.prototype._sortCombatants = _sortCombatants;
+  Combat.prototype.resetAll = resetAll;
   MeasuredTemplate.prototype._getConeShape = getSwadeConeShape;
 
   // Register custom classes